+++ R/calc_size.R
@@ -23,25 +23,13 @@ calc_size <- function(pop_n, pop_mean_size, pop_linf, pop_var_size,
   # use log distribution for starting size
   if (use_log) {
 
-    # calc mean of log-norm distribution
-    norm_mean <- log((pop_mean_size ^ 2) / sqrt(pop_mean_size ^ 2 + pop_var_size))
-
-    # calc sd of log-norm distribution
-    norm_sd <- sqrt(log(1 + (pop_var_size / (pop_mean_size ^ 2))))
-
-
-    # get random numbers from log-norm distribution
-    norm_random <- stats::rnorm(n = pop_n,
-                                mean = norm_mean, sd = norm_sd)
-
-    # calculate body length based on random number
-    body_length <- exp(norm_random)
+    body_length <- rlognorm(n = pop_n, mean = pop_mean_size,
+                            sd = sqrt(pop_var_size))
 
   # use uniform distribution for starting size
   } else {
 
-    body_length <- stats::runif(n = pop_n,
-                                min = pop_linf * 0.1, max = pop_linf * 0.9)
+    body_length <- stats::runif(n = pop_n, min = pop_linf * 0.1, max = pop_linf * 0.9)
 
   }